+++ crates/teloxide-core/src/adaptors/erased.rs
@@ -200,6 +200,7 @@ where
         copy_message,
         copy_messages,
         send_message,
+        send_message_draft,
         send_photo,
         send_audio,
         send_document,
@@ -335,9 +336,12 @@ where
         get_business_account_star_balance,
         transfer_business_account_stars,
         get_business_account_gifts,
+        get_user_gifts,
+        get_chat_gifts,
         convert_gift_to_stars,
         upgrade_gift,
         transfer_gift,
+        repost_story,
         post_story,
         edit_story,
         delete_story,
@@ -384,6 +388,13 @@ trait ErasableRequester<'a> {
         chat_id: Recipient,
         text: String,
     ) -> ErasedRequest<'a, SendMessage, Self::Err>;
+
+    fn send_message_draft(
+        &self,
+        chat_id: ChatId,
+        draft_id: i64,
+        text: String,
+    ) -> ErasedRequest<'a, SendMessageDraft, Self::Err>;
 
     fn forward_message(
         &self,
@@ -1140,6 +1151,10 @@ trait ErasableRequester<'a> {
         business_connection_id: BusinessConnectionId,
     ) -> ErasedRequest<'a, GetBusinessAccountGifts, Self::Err>;
 
+    fn get_user_gifts(&self, user_id: UserId) -> ErasedRequest<'a, GetUserGifts, Self::Err>;
+
+    fn get_chat_gifts(&self, chat_id: Recipient) -> ErasedRequest<'a, GetChatGifts, Self::Err>;
+
     fn convert_gift_to_stars(
         &self,
         business_connection_id: BusinessConnectionId,
@@ -1158,6 +1173,14 @@ trait ErasableRequester<'a> {
         owned_gift_id: OwnedGiftId,
         new_owner_chat_id: ChatId,
     ) -> ErasedRequest<'a, TransferGift, Self::Err>;
+
+    fn repost_story(
+        &self,
+        business_connection_id: BusinessConnectionId,
+        from_chat_id: Recipient,
+        from_story_id: StoryId,
+        active_period: Seconds,
+    ) -> ErasedRequest<'a, RepostStory, Self::Err>;
 
     fn post_story(
         &self,
@@ -1306,6 +1329,15 @@ where
         text: String,
     ) -> ErasedRequest<'a, SendMessage, Self::Err> {
         Requester::send_message(self, chat_id, text).erase()
+    }
+
+    fn send_message_draft(
+        &self,
+        chat_id: ChatId,
+        draft_id: i64,
+        text: String,
+    ) -> ErasedRequest<'a, SendMessageDraft, Self::Err> {
+        Requester::send_message_draft(self, chat_id, draft_id, text).erase()
     }
 
     fn forward_message(
@@ -2374,6 +2406,14 @@ where
         Requester::get_business_account_gifts(self, business_connection_id).erase()
     }
 
+    fn get_user_gifts(&self, user_id: UserId) -> ErasedRequest<'a, GetUserGifts, Self::Err> {
+        Requester::get_user_gifts(self, user_id).erase()
+    }
+
+    fn get_chat_gifts(&self, chat_id: Recipient) -> ErasedRequest<'a, GetChatGifts, Self::Err> {
+        Requester::get_chat_gifts(self, chat_id).erase()
+    }
+
     fn convert_gift_to_stars(
         &self,
         business_connection_id: BusinessConnectionId,
@@ -2398,6 +2438,23 @@ where
     ) -> ErasedRequest<'a, TransferGift, Self::Err> {
         Requester::transfer_gift(self, business_connection_id, owned_gift_id, new_owner_chat_id)
             .erase()
+    }
+
+    fn repost_story(
+        &self,
+        business_connection_id: BusinessConnectionId,
+        from_chat_id: Recipient,
+        from_story_id: StoryId,
+        active_period: Seconds,
+    ) -> ErasedRequest<'a, RepostStory, Self::Err> {
+        Requester::repost_story(
+            self,
+            business_connection_id,
+            from_chat_id,
+            from_story_id,
+            active_period,
+        )
+        .erase()
     }
 
     fn post_story(