+++ crates/teloxide-core/src/types/gift_info.rs
@@ -36,6 +36,16 @@ pub struct GiftInfo {
     #[serde(default, skip_serializing_if = "std::ops::Not::not")]
     pub can_be_upgraded: bool,
 
+    /// `true`, if the gift can only be upgraded on its own, and the option to
+    /// upgrade the gift and add more Telegram Stars to the upgrade
+    /// simultaneously isn't available
+    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
+    pub is_upgrade_separate: bool,
+
+    /// Number of the corresponding unique gift, if the gift is upgraded and the
+    /// number is known
+    pub unique_gift_number: Option<u64>,
+
     /// Text of the message that was added to the gift
     pub text: Option<String>,