+++ crates/teloxide-core/src/types/user.rs
@@ -36,6 +36,11 @@ pub struct User {
     /// `true`, if this user added the bot to the attachment menu.
     #[serde(default, skip_serializing_if = "std::ops::Not::not")]
     pub added_to_attachment_menu: bool,
+
+    /// `true`, if forum topic mode is enabled for the bot in private chats with
+    /// the user.
+    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
+    pub has_topics_enabled: bool,
 }
 
 impl User {
@@ -152,6 +157,7 @@ mod tests {
             language_code: Some(String::from("ru")),
             is_premium: false,
             added_to_attachment_menu: false,
+            has_topics_enabled: false,
         };
         let actual = serde_json::from_str::<User>(json).unwrap();
         assert_eq!(actual, expected)
@@ -168,6 +174,7 @@ mod tests {
             language_code: None,
             is_premium: false,
             added_to_attachment_menu: false,
+            has_topics_enabled: false,
         };
 
         let user_b = User {
@@ -179,6 +186,7 @@ mod tests {
             language_code: None,
             is_premium: false,
             added_to_attachment_menu: false,
+            has_topics_enabled: false,
         };
 
         assert_eq!(user_a.full_name(), "First Last");